requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
twilio==9.0.4
openai==1.3.7
tiktoken==0.5.2
pandas==2.1.4
numpy==1.24.3
networkx==3.2.1
sentence-transformers==2.2.2
nltk==3.8.1
textstat==0.7.3
aiofiles==23.2.1
python-multipart==0.0.6